updateMerchant and changed vat from double to integer since i think its supposed to be an ID

+++ src/main/java/com/example/wdmsystem/merchant/system/Merchant.java
@@ -14,7 +14,7 @@ public class Merchant {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     public Integer id;
     public String name;
-    public double vat;
+    public Integer vat; // i think this is supposed to be a tax id and not the vat itself
     @Embedded
     public Address address;
     public String email;
@@ -22,7 +22,7 @@ public class Merchant {
     public LocalDateTime createdAt;
     public LocalDateTime updatedAt;
 
-    public Merchant(Integer id, String name, double vat, Address address, String email, String phone, LocalDateTime createdAt) {
+    public Merchant(Integer id, String name, Integer vat, Address address, String email, String phone, LocalDateTime createdAt) {
         this.id = id;
         this.name = name;
         this.vat = vat;
